Refactor: StatusItem 컴포넌트 분리, divide-x 활용
